MATE-63: [REFACTOR] 굿즈거래 판매글 삭제 로직 수정 (#40)
* MATE-63: [REFACTOR] 굿즈거래 판매글 삭제 로직 수정

- 굿즈 판매글을 삭제할 때 거래완료인 상태에는 예외를 발생시키도록 수정

* MATE-63: [TEST] 굿즈거래 판매글 삭제 테스트 시나리오 추가

+++ src/main/java/com/example/mate/domain/goods/service/GoodsService.java
@@ -75,6 +75,10 @@ public class GoodsService {
         Member seller = getSellerAndValidate(memberId);
         GoodsPost goodsPost = getGoodsPostAndValidate(seller, goodsPostId);
 
+        if (goodsPost.getStatus() == Status.CLOSED) {
+            throw new CustomException(ErrorCode.GOODS_DELETE_NOT_ALLOWED);
+        }
+
         // 업로된 이미지 파일 삭제
         deleteExistingImages(goodsPostId);
         goodsPostRepository.delete(goodsPost);
